Move SlashCommand Generator to Command class, Re-implement dynamic module reload
